Struct google_gameservices1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the GameServices hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_gameservices1 as gameservices1;
 
use std::default::Default;
use oauth2;
use gameservices1::GameServices;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = GameServices::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_game_server_deployments_configs_create(...)`, `locations_game_server_deployments_configs_delete(...)`, `locations_game_server_deployments_configs_get(...)`, `locations_game_server_deployments_configs_list(...)`, `locations_game_server_deployments_create(...)`, `locations_game_server_deployments_delete(...)`, `locations_game_server_deployments_fetch_deployment_state(...)`, `locations_game_server_deployments_get(...)`, `locations_game_server_deployments_get_iam_policy(...)`, `locations_game_server_deployments_get_rollout(...)`, `locations_game_server_deployments_list(...)`, `locations_game_server_deployments_patch(...)`, `locations_game_server_deployments_preview_rollout(...)`, `locations_game_server_deployments_set_iam_policy(...)`, `locations_game_server_deployments_test_iam_permissions(...)`, `locations_game_server_deployments_update_rollout(...)`, `locations_get(...)`, `locations_list(...)`, `locations_operations_cancel(...)`, `locations_operations_delete(...)`, `locations_operations_get(...)`, `locations_operations_list(...)`, `locations_realms_create(...)`, `locations_realms_delete(...)`, `locations_realms_game_server_clusters_create(...)`, `locations_realms_game_server_clusters_delete(...)`, `locations_realms_game_server_clusters_get(...)`, `locations_realms_game_server_clusters_list(...)`, `locations_realms_game_server_clusters_patch(...)`, `locations_realms_game_server_clusters_preview_create(...)`, `locations_realms_game_server_clusters_preview_delete(...)`, `locations_realms_game_server_clusters_preview_update(...)`, `locations_realms_get(...)`, `locations_realms_list(...)`, `locations_realms_patch(...)` and `locations_realms_preview_update(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_game_server_deployments_configs_create(
&self,
request: GameServerConfig,
parent: &str
) -> ProjectLocationGameServerDeploymentConfigCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new game server config in a given project, location, and game server deployment. Game server configs are immutable, and are not applied until referenced in the game server deployment rollout resource.

Arguments

request - No description provided.

parent - Required. The parent resource name, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}/.

pub fn locations_game_server_deployments_configs_delete(
&self,
name: &str
) -> ProjectLocationGameServerDeploymentConfigDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single game server config. The deletion will fail if the game server config is referenced in a game server deployment rollout.

Arguments

name - Required. The name of the game server config to delete, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}/configs/{config}.

pub fn locations_game_server_deployments_configs_get(
&self,
name: &str
) -> ProjectLocationGameServerDeploymentConfigGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single game server config.

Arguments

name - Required. The name of the game server config to retrieve, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}/configs/{config}.

pub fn locations_game_server_deployments_configs_list(
&self,
parent: &str
) -> ProjectLocationGameServerDeploymentConfigListCall<'a, C>

Create a builder to help you perform the following task:

Lists game server configs in a given project, location, and game server deployment.

Arguments

parent - Required. The parent resource name, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}/configs/*.

pub fn locations_game_server_deployments_create(
&self,
request: GameServerDeployment,
parent: &str
) -> ProjectLocationGameServerDeploymentCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new game server deployment in a given project and location.

Arguments

request - No description provided.

parent - Required. The parent resource name, in the following form: projects/{project}/locations/{location}.

pub fn locations_game_server_deployments_delete(
&self,
name: &str
) -> ProjectLocationGameServerDeploymentDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single game server deployment.

Arguments

name - Required. The name of the game server delpoyment to delete, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}.

pub fn locations_game_server_deployments_fetch_deployment_state(
&self,
request: FetchDeploymentStateRequest,
name: &str
) -> ProjectLocationGameServerDeploymentFetchDeploymentStateCall<'a, C>

Create a builder to help you perform the following task:

Retrieves information about the current state of the game server deployment. Gathers all the Agones fleets and Agones autoscalers, including fleets running an older version of the game server deployment.

Arguments

request - No description provided.

name - Required. The name of the game server delpoyment, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}.

pub fn locations_game_server_deployments_get(
&self,
name: &str
) -> ProjectLocationGameServerDeploymentGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single game server deployment.

Arguments

name - Required. The name of the game server delpoyment to retrieve, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}.

pub fn locations_game_server_deployments_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationGameServerDeploymentGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_game_server_deployments_get_rollout(
&self,
name: &str
) -> ProjectLocationGameServerDeploymentGetRolloutCall<'a, C>

Create a builder to help you perform the following task:

Gets details a single game server deployment rollout.

Arguments

name - Required. The name of the game server delpoyment to retrieve, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}/rollout.

pub fn locations_game_server_deployments_list(
&self,
parent: &str
) -> ProjectLocationGameServerDeploymentListCall<'a, C>

Create a builder to help you perform the following task:

Lists game server deployments in a given project and location.

Arguments

parent - Required. The parent resource name, in the following form: projects/{project}/locations/{location}.

pub fn locations_game_server_deployments_patch(
&self,
request: GameServerDeployment,
name: &str
) -> ProjectLocationGameServerDeploymentPatchCall<'a, C>

Create a builder to help you perform the following task:

Patches a game server deployment.

Arguments

request - No description provided.

name - The resource name of the game server deployment, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}. For example, projects/my-project/locations/global/gameServerDeployments/my-deployment.

pub fn locations_game_server_deployments_preview_rollout(
&self,
request: GameServerDeploymentRollout,
name: &str
) -> ProjectLocationGameServerDeploymentPreviewRolloutCall<'a, C>

Create a builder to help you perform the following task:

Previews the game server deployment rollout. This API does not mutate the rollout resource.

Arguments

request - No description provided.

name - The resource name of the game server deployment rollout, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}/rollout. For example, projects/my-project/locations/global/gameServerDeployments/my-deployment/rollout.

pub fn locations_game_server_deployments_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationGameServerDeploymentSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_game_server_deployments_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationGameServerDeploymentTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_game_server_deployments_update_rollout(
&self,
request: GameServerDeploymentRollout,
name: &str
) -> ProjectLocationGameServerDeploymentUpdateRolloutCall<'a, C>

Create a builder to help you perform the following task:

Patches a single game server deployment rollout. The method will not return an error if the update does not affect any existing realms. For example - if the default_game_server_config is changed but all existing realms use the override, that is valid. Similarly, if a non existing realm is explicitly called out in game_server_config_overrides field, that will also not result in an error.

Arguments

request - No description provided.

name - The resource name of the game server deployment rollout, in the following form: projects/{project}/locations/{location}/gameServerDeployments/{deployment}/rollout. For example, projects/my-project/locations/global/gameServerDeployments/my-deployment/rollout.

pub fn locations_operations_cancel(
&self,
request: CancelOperationRequest,
name: &str
) -> ProjectLocationOperationCancelCall<'a, C>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

request - No description provided.

name - The name of the operation resource to be cancelled.

pub fn locations_operations_delete(
&self,
name: &str
) -> ProjectLocationOperationDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_operations_list(
&self,
name: &str
) -> ProjectLocationOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_realms_game_server_clusters_create(
&self,
request: GameServerCluster,
parent: &str
) -> ProjectLocationRealmGameServerClusterCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new game server cluster in a given project and location.

Arguments

request - No description provided.

parent - Required. The parent resource name, in the following form: projects/{project}/locations/{location}/realms/{realm-id}.

pub fn locations_realms_game_server_clusters_delete(
&self,
name: &str
) -> ProjectLocationRealmGameServerClusterDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single game server cluster.

Arguments

name - Required. The name of the game server cluster to delete, in the following form: projects/{project}/locations/{location}/gameServerClusters/{cluster}.

pub fn locations_realms_game_server_clusters_get(
&self,
name: &str
) -> ProjectLocationRealmGameServerClusterGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single game server cluster.

Arguments

name - Required. The name of the game server cluster to retrieve, in the following form: projects/{project}/locations/{location}/realms/{realm-id}/gameServerClusters/{cluster}.

pub fn locations_realms_game_server_clusters_list(
&self,
parent: &str
) -> ProjectLocationRealmGameServerClusterListCall<'a, C>

Create a builder to help you perform the following task:

Lists game server clusters in a given project and location.

Arguments

parent - Required. The parent resource name, in the following form: "projects/{project}/locations/{location}/realms/{realm}".

pub fn locations_realms_game_server_clusters_patch(
&self,
request: GameServerCluster,
name: &str
) -> ProjectLocationRealmGameServerClusterPatchCall<'a, C>

Create a builder to help you perform the following task:

Patches a single game server cluster.

Arguments

request - No description provided.

name - Required. The resource name of the game server cluster, in the following form: projects/{project}/locations/{location}/realms/{realm}/gameServerClusters/{cluster}. For example, projects/my-project/locations/{location}/realms/zanzibar/gameServerClusters/my-onprem-cluster.

pub fn locations_realms_game_server_clusters_preview_create(
&self,
request: GameServerCluster,
parent: &str
) -> ProjectLocationRealmGameServerClusterPreviewCreateCall<'a, C>

Create a builder to help you perform the following task:

Previews creation of a new game server cluster in a given project and location.

Arguments

request - No description provided.

parent - Required. The parent resource name, in the following form: projects/{project}/locations/{location}/realms/{realm}.

pub fn locations_realms_game_server_clusters_preview_delete(
&self,
name: &str
) -> ProjectLocationRealmGameServerClusterPreviewDeleteCall<'a, C>

Create a builder to help you perform the following task:

Previews deletion of a single game server cluster.

Arguments

name - Required. The name of the game server cluster to delete, in the following form: projects/{project}/locations/{location}/gameServerClusters/{cluster}.

pub fn locations_realms_game_server_clusters_preview_update(
&self,
request: GameServerCluster,
name: &str
) -> ProjectLocationRealmGameServerClusterPreviewUpdateCall<'a, C>

Create a builder to help you perform the following task:

Previews updating a GameServerCluster.

Arguments

request - No description provided.

name - Required. The resource name of the game server cluster, in the following form: projects/{project}/locations/{location}/realms/{realm}/gameServerClusters/{cluster}. For example, projects/my-project/locations/{location}/realms/zanzibar/gameServerClusters/my-onprem-cluster.

pub fn locations_realms_create(
&self,
request: Realm,
parent: &str
) -> ProjectLocationRealmCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new realm in a given project and location.

Arguments

request - No description provided.

parent - Required. The parent resource name, in the following form: projects/{project}/locations/{location}.

pub fn locations_realms_delete(
&self,
name: &str
) -> ProjectLocationRealmDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single realm.

Arguments

name - Required. The name of the realm to delete, in the following form: projects/{project}/locations/{location}/realms/{realm}.

pub fn locations_realms_get(
&self,
name: &str
) -> ProjectLocationRealmGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single realm.

Arguments

name - Required. The name of the realm to retrieve, in the following form: projects/{project}/locations/{location}/realms/{realm}.

pub fn locations_realms_list(
&self,
parent: &str
) -> ProjectLocationRealmListCall<'a, C>

Create a builder to help you perform the following task:

Lists realms in a given project and location.

Arguments

parent - Required. The parent resource name, in the following form: projects/{project}/locations/{location}.

pub fn locations_realms_patch(
&self,
request: Realm,
name: &str
) -> ProjectLocationRealmPatchCall<'a, C>

Create a builder to help you perform the following task:

Patches a single realm.

Arguments

request - No description provided.

name - The resource name of the realm, in the following form: projects/{project}/locations/{location}/realms/{realm}. For example, projects/my-project/locations/{location}/realms/my-realm.

pub fn locations_realms_preview_update(
&self,
request: Realm,
name: &str
) -> ProjectLocationRealmPreviewUpdateCall<'a, C>

Create a builder to help you perform the following task:

Previews patches to a single realm.

Arguments

request - No description provided.

name - The resource name of the realm, in the following form: projects/{project}/locations/{location}/realms/{realm}. For example, projects/my-project/locations/{location}/realms/my-realm.

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.